requirements.txt
sentence_transformers==2.2.0
nltk==3.6.5
numpy==1.21.6
pandas==1.3.5
spacy==3.2.4
tqdm==4.64.0
openpyxl==3.0.10
pymongo[srv]==4.2.0
streamlit==1.13.0
faiss-cpu==1.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl